Code View and Code Edit tweaked
The Code Edit page can now submit title, content, language, and code all
to the database. The Code View page can display all of the
aforementioned information. The only issue is that the getting the
language of the code viewer to change to the type of language. Getting
$row[3] to turn into a JS variable for the
editor.session.setMode("ace/mode/(language goes here)"); FIX THIS
SOMEONE